Separated GPU resource module from State
Introduces a new module GpuResource that hold the communication with the GPU indepndent from the reaction system. Surface, Device, queue, config and PhysicalSize of the window are defined here and created. It also introduces resize (for when the user resize the window), begin_frame which gets the texture and creates view and command encoder. Moreover, it contains submit_frame function where the queue is submitted.