Added ping-pong texture
Introduces the ping pong texture (not the compure still). Once the GPU is reading from one texture the results are being written in another one and then the GPU switches. In this way we make sure that the data is not corrupted and the we are seeing the correct data on the correct frame. In this commit a circular shaped blob with a radius of 10 pixels is also added for the reaction to start in the middle of the screen.