Added ping-pong compute
Introduced the two know bindings to the compute BGL for the texture source and destination texture (write only). Now the there are two BG for compute. At this stage program runs with still showing the blue screen. We still need to add rendering and the vertex/fragment shaders for the diffusion to become visible