Added shader watcher for hot reload
Introduces the core code for shader watcher that checks during the runtime if there was any change in the shaders and if yes apply it to the code. In this way we can change the color or parameters while the code is running and the changes are applied without any need to rerun the code.