Fixed hardcoded brush radius
There was also another issue where a blob was at the bottom left corner that got changed in size whenever I changed the size of the brush. I had to make an if statement for the brush size so it stays inactive when when the left click is not down. However the blob didn't go away and I had to change something in the brush compute shader