Added debuge mode
The debug mode is wired to per frame parameters (in node.rs) and can be activated using key D. While the program is running the debug mode can be acticvated which is meant to help with knowing if the nodes are being executed and if the order is correct. Of course as of now I just added one Node and that is the ReactionDiffusionNode. I will expand it and I want to separate an 'Injection Node' for mouse inputes later...