Wired RenderGraph to State
Instead of accessing the ReactionDiffusionSystem directly the program now goes through a RenderGraph. The RenderGraph then can execute the compute and render pass. I was also able to take out the brush parameters when mouse is clicked on the screen. Those parameters now live in ReactionDiffusionNode and not directly in the State.