Introducing RenderNode
Here PerFrameParameters live as well as the trait RenderNode which defines the contract for every node. For now every node needs to have a name for debugging, they should be able to be executed, the should allow hot reload (later I want to add post processing and some filters that must also allow for hot reload just like the ReactionDiffusion itself). Also they must allow for downcasting so that we can access some of the funcions.