Introducing RenderGraph
The render graph needs the resource reginstry (later I need to add them. Right now it is not being really used) and the nodes. Nodes are able to be added via add_node, and executed via execute that call the execute function from the node. Notifying the hotreload add well as downcasting are also defined here. One allows for hotreload for all the nodes and the downcasting is needed to get access to some functionalities.