Added RD output aware registry
The registry doesn't really do anything in this step. I just added it as a checkpoint before giving it the ownership of some textures. The idea is that the registry must enable the nodes to communicate via names and not refrences. The shared resources must then also become exposed via registry.