Added necessary compute registry functions and updated API
changed the name of the function to match them to their responsibilities more accurately. For example instead of register_texture that was meant for me to know where the function lives (in this example in registry) I change the name of that function to get_texture because that is what the function do. it gets the name. For the compute I also added storage_texture_creator function which can create texture and store it.