Added PassType for Graph to order nodes
Until now I had to order the nodes when using add_node manually. But now after adding enum PassType with correct ordering the graph can do the ordering itself. Right now we have only two nodes 'compute' and 'render' and compute comes before render. When writing a new node, we have to give it a type. With the help of this type the graph knows the order.