Introduced Substeps for stability
The solver has now SimulationParameters where I introduced dt and substeps. The dt is not hardcoded like before anymore. With the help of the substeps we can also now make the rendered solutions with bigger steps thus we see the simulation is running faster without losing stability in compute. Also I had to break down the compute function and take the mouse injection out to make sure that the injection is not being applied in every unrendered frame. I believe with the new code the mouse injection should be applied in correct frame.